% Generated by roxygen2 (4.0.0): do not edit by hand
\name{author.coincidence.graph}
\alias{author.coincidence.graph}
\title{Create a graph from conversation data}
\usage{
author.coincidence.graph(conversation, author,
  similarity.measure = "coincidence_count")
}
\arguments{
\item{conversation.id}{A vector representing the conversation authors were engaged in (e.g. a thread, meeting)}

\item{author}{A vector with unique author names/ids}
}
\value{
A graph object in the \code{\link{igraph}} format
}
\description{
Create a graph from conversation data
}

