% Generated by roxygen2 (4.0.0): do not edit by hand
\name{content.similarity.graph}
\alias{content.similarity.graph}
\title{Create a content similarity graph}
\usage{
content.similarity.graph(m, vertex.grouping.vars,
  similarity.measure = "cosine", topicscores.as.vertexmeta = NULL,
  min.similarity = NULL)
}
\arguments{
\item{m}{A (sparse) matrix where rows are documents (e.g., news articles, forum posts) and columns are content characteristics (e.g., terms, topics). Values represent the presence of content characteristics within documents. Examples are: a \code{\link{DocumentTermMatrix}} or the transposed $document_sums (topics by documents matrix) created with \code{\link{lda.collapsed.gibbs.sampler}}.}

\item{vertex.grouping.vars}{A data.frame or list with named vectors representing vertex characteristics. Each unique combination of characteristics will be considered a vertex. In the graph object these characteristics are stored as vertex attributes.}

\item{similarity.measure}{A character string giving a method for computing similarity. Options are: 'correlation', 'cosine','conditional_probability','overlap_count' and 'overlap_jacard'.}

\item{topicscores.as.vertexmeta}{Can be used to include the topic values per node per topic. if 'all', all topics will be included. Can also be a numeric vector to select specific topics.}

\item{min.similarity}{A numeric scalar representing the threshold for similarities. All ties with a value below min.similarity will be deleted. Can be used to reduce the size of large graphs with many weak ties.}
}
\value{
A graph object in the \code{\link{igraph}} format
}
\description{
Makes a document similarity graph in the \code{\link{igraph}} format, based on a matrix in which rows are documents, and columns are content characteristics (e.g., terms, issues, topics).
Vertices (i.e. nodes) are defined as unique combinations of vertex.grouping.vars. If vertices cover multiple documents (e.g., authors of several documents) then content characteristics are first aggregated.
The (aggregated) content characteristics are used to calculate the similarities between vertices, for which various similarity measures can be used
}

