% Generated by roxygen2 (4.0.0): do not edit by hand
\name{conversation.graph.create}
\alias{conversation.graph.create}
\title{Create a graph from conversation data}
\usage{
conversation.graph.create(conversation, author, order.nr = NULL,
  order.lookback = NULL, direction = "undirected")
}
\arguments{
\item{conversation.id}{A vector representing the conversation authors were engaged in (e.g. a thread, meeting)}

\item{author}{A vector with unique author names/ids}

\item{order.nr}{Optional. A vector representing the order in which authors communicated within a conversation.}

\item{order.lookback}{Requires an order.nr vector. If value is 'first', a tie will be drawn between each author with the first author of a converstation. If value is a (non-negative) numerical scalar (x), ties are drawn between authors with the previous x authors. For example, if value is 1, ties are drawn between an author and the previous author. If value is 2, between an author and the previous 2 authors. etc.}

\item{direction}{A character string, indicating the direction of the ties. If 'undirected', ties between authors go both ways. If 'directed.up', ties are directed from author to previous authors (only if order is given). if 'directed.down', ties are directed from author to later authors.}
}
\value{
A data.frame with the columns id, order and substring
}
\description{
Create a graph from conversation data
}

