% Generated by roxygen2 (4.0.0): do not edit by hand
\name{default.graph.attributes}
\alias{default.graph.attributes}
\title{Sets basic default values for content.similarity.graph}
\usage{
default.graph.attributes(g)
}
\arguments{
\item{g}{A graph object in the \code{\link{igraph}} format}
}
\value{
A graph object in the \code{\link{igraph}} format
}
\description{
Sets basic default values for content.similarity.graph
}

