% Generated by roxygen2 (4.0.0): do not edit by hand
\name{graph.color.vertices}
\alias{graph.color.vertices}
\title{Assign vertex colors to a graph}
\usage{
graph.color.vertices(g, attribute, pallete = NULL)
}
\arguments{
\item{g}{A graph object in the \code{\link{igraph}} format}

\item{attribute}{A vertex attribute to organize the colors. Either use a categorical vector (factor, character) to color categories in different colors, or use a numerical vector (numerical, integer) to highlight with one color (like a heat map)}

\item{pallete}{Optional. Add a custom pallete, which should be a character vector with color codes that igraph can use.}
}
\value{
A graph object in the \code{\link{igraph}} format
}
\description{
A convenience function to add colors to a graph of the \code{\link{igraph}} format based on a vector of vertex attributes.
}

