% Generated by roxygen2 (4.0.0): do not edit by hand
\name{matrix.aggregate}
\alias{matrix.aggregate}
\title{Aggregates a (sparse) matrix}
\usage{
matrix.aggregate(mat, grouping.vars, FUN = "sum")
}
\arguments{
\item{mat}{A matrix object, can be a sparse matrix}

\item{grouping.vars}{The grouping elements. Can be a single vector, or a list or data.frame with multiple vectors}

\item{The}{name of the function to be applied}
}
\value{
A list containing a (sparse) matrix and a data.frame with the variables matching the matrix rows
}
\description{
Uses the \code{\link{slam}} package rollup function to efficiently aggregate a sparse matrix. Also allows multiple vectors to be used as grouping vars.
}

