% Generated by roxygen2 (4.0.0): do not edit by hand
\name{matrix.coincidence.count}
\alias{matrix.coincidence.count}
\title{Compute coincidence count for matrix columns}
\usage{
matrix.coincidence.count(mat)
}
\arguments{
\item{m}{A matrix object. Can be a sparse matrix.}
}
\value{
An adjacency matrix
}
\description{
Computes the coincidence of all columnso of a matrix
}

