% Generated by roxygen2 (4.0.0): do not edit by hand
\name{matrix.conprob}
\alias{matrix.conprob}
\title{Compute conditional probability for matrix columns}
\usage{
matrix.conprob(mat, alpha = 2)
}
\arguments{
\item{mat}{A matrix object. Can be a sparse matrix.}

\item{alpha}{A weighting value to transform count to probabilities}
}
\value{
An adjacency matrix
}
\description{
Computes the conditional probability for all columns of a matrix
}

