% Generated by roxygen2 (4.0.0): do not edit by hand
\name{matrix.correlation}
\alias{matrix.correlation}
\title{Compute correlations for matrix columns}
\usage{
matrix.correlation(mat)
}
\arguments{
\item{m}{A matrix object. Can be a sparse matrix.}
}
\value{
An adjacency matrix
}
\description{
Computes correlations between all columnso of a matrix
}

