% Generated by roxygen2 (4.0.0): do not edit by hand
\name{matrix.cosine}
\alias{matrix.cosine}
\title{Compute cosine for matrix columns}
\usage{
matrix.cosine(mat)
}
\arguments{
\item{m}{A matrix object. Can be a sparse matrix.}
}
\value{
An adjacency matrix
}
\description{
Computes the cosine distance for all columns of a matrix
}

