% Generated by roxygen2 (4.0.0): do not edit by hand
\name{matrix.overlap.jacard}
\alias{matrix.overlap.jacard}
\title{Compute jacard for matrix columns}
\usage{
matrix.overlap.jacard(mat)
}
\arguments{
\item{m}{A matrix object. Can be a sparse matrix.}
}
\value{
An adjacency matrix
}
\description{
Computes jacard between all columnso of a matrix
}

