% Generated by roxygen2 (4.0.0): do not edit by hand
\name{previous.authors.graph}
\alias{previous.authors.graph}
\title{Create a graph from conversation data}
\usage{
previous.authors.graph(conversation, author, order.nr, lookback, direction)
}
\arguments{
\item{conversation.id}{A vector representing the conversation authors were engaged in (e.g. a thread, meeting)}

\item{author}{A vector with unique author names/ids}

\item{order.nr}{Optional. A vector representing the order in which authors communicated within a conversation.}

\item{lookback}{The number of previous author with whom a tie with an author should be drawn}

\item{direction}{A character string, indicating the direction of the ties. If 'undirected', ties between authors go both ways. If 'directed.up', ties are directed from author to previous authors (only if order is given). if 'directed.down', ties are directed from author to later authors.}
}
\value{
A graph object in the \code{\link{igraph}} format
}
\description{
Create a graph from conversation data, in which ties are drawn between authors and the authors that communicated within a given distance (lookback) before them.
}

