% Generated by roxygen2 (4.0.0): do not edit by hand
\name{splitstr.to.rows}
\alias{splitstr.to.rows}
\title{Split strings to rows with matching ids}
\usage{
splitstr.to.rows(x, split_by = ";", id = NULL, ...)
}
\arguments{
\item{x}{A character vector.}

\item{split_by}{The character vector by which x will be split. By default should be a regular expression (see \code{\link{strsplit}}) but can be changed using the ellipsis parameters}

\item{id}{The id to be assigned to the rows. If a vector of the same lengt of x is given, these values will be used.}

\item{...}{additional parameters to be passed to the strsplit function}
}
\value{
A data.frame with the columns id, order and substring
}
\description{
A convenience function. Mainly designed for preparing data for co-author type networks, if an unknown number of authors per document is stored as a single character string with delimiters.
A character string is split by a given character vector. For each part of the string a new row is created. An id value indicates which parts belong together. An order value indicates which part of the original string the stringpart was
}

