% Generated by roxygen2 (4.0.0): do not edit by hand
\name{first.author.graph}
\alias{first.author.graph}
\title{Create a graph from conversation data}
\usage{
first.author.graph(conversation, author, order.nr, direction = "directed.up",
  once.per.conversation = TRUE)
}
\arguments{
\item{conversation.id}{A vector representing the conversation authors were engaged in (e.g. a thread, meeting)}

\item{author}{A vector with unique author names/ids}

\item{order.nr}{Optional. A vector representing the order in which authors communicated within a conversation.}

\item{direction}{A character string, indicating the direction of the ties. If 'undirected', ties between authors go both ways. If 'directed.up', ties are directed from author to previous authors (only if order is given). if 'directed.down', ties are directed from author to later authors.}

\item{once.per.conversation}{Logical. If TRUE, count only the number of conversations in which a first-author <-> later-author coincidence occurs. It is then ignored how many times the later author participated in the conversation. If FALSE, the number of participation times is counted.}
}
\value{
A graph object in the \code{\link{igraph}} format
}
\description{
Create a graph from conversation data, in which ties are drawn between each authors in the conversation and the first author.
}

