% Generated by roxygen2 (4.0.0): do not edit by hand
\name{matrix.rowsimilarities}
\alias{matrix.rowsimilarities}
\title{Calculates the similarity between rows of a matrix}
\usage{
matrix.rowsimilarities(mat, similarity.measure, output.as = "matrix",
  min.similarity = NULL, decimals = 3)
}
\arguments{
\item{matrix}{A matrix object, which can be a sparse matrix.}

\item{similarity.measure}{A character string giving a method for computing similarity. Options are: 'correlation', 'cosine','conditional_probability','coincidence_count' and 'overlap_jacard'.}

\item{output}{A character string indicating whether output should be a 'matrix' or a 'graph'. The graph is directed if assymetrical similarity measures are used ('conditional_probability' or 'overlap_jacard')}
}
\value{
An adjacency matrix or graph object
}
\description{
Calculates the similarity between rows of a matrix using various similarity/distance measures. Output can be a matrix or a graph object in the \code{\link{igraph}} format
}

